Agrega componente NavBar  ylo agrega en las pages de Login y Menu.
