+++ src/components/NewOrder.jsx
@@ -1,0 +1,5 @@
+const NewOrder = () => {
+	return <div></div>;
+};
+
+export default NewOrder;
