+++ src/main.jsx
@@ -5,12 +5,12 @@ import {
 	BrowserRouter
   } from "react-router-dom";
 import './index.css';
+import { NavBar } from './components/NavBar';
 
 ReactDOM.render(
-	
-		<BrowserRouter>
-			<App />
-		</BrowserRouter>,
-	
+	<BrowserRouter>
+		<App />
+	</BrowserRouter>,
+
 	document.getElementById('root')
 );
